+++ packages/core/src/config/sources.ts
