+++ packages/core/src/sync-historical/service.ts
@@ -20,7 +20,7 @@ import {
   type LogFilterCriteria,
   type Source,
   sourceIsLogFilter,
-} from "@/config/types.js";
+} from "@/config/sources.js";
 import { getHistoricalSyncStats } from "@/metrics/utils.js";
 import type { SyncStore } from "@/sync-store/store.js";
 import type { Checkpoint } from "@/utils/checkpoint.js";
