+++ packages/core/src/sync-realtime/bloom.ts
@@ -4,7 +4,7 @@ import {
 } from "ethereum-bloom-filters";
 import type { Address, Hex } from "viem";
 
-import type { Topics } from "@/config/types.js";
+import type { Topics } from "@/config/sources.js";
 
 export function isMatchedLogInBloomFilter({
   bloom,
