+++ packages/core/src/sync-store/sqlite/store.ts
@@ -15,7 +15,7 @@ import {
 } from "viem";
 
 import type { Common } from "@/Ponder.js";
-import type { FactoryCriteria, LogFilterCriteria } from "@/config/types.js";
+import type { FactoryCriteria, LogFilterCriteria } from "@/config/sources.js";
 import type { Block } from "@/types/block.js";
 import type { Log } from "@/types/log.js";
 import type { Transaction } from "@/types/transaction.js";
