+++ packages/core/src/ui/service.ts
@@ -1,5 +1,5 @@
 import type { Common } from "@/Ponder.js";
-import type { Source } from "@/config/types.js";
+import type { Source } from "@/config/sources.js";
 import { getHistoricalSyncStats } from "@/metrics/utils.js";
 
 import { type UiState, buildUiState, setupInkApp } from "./app.js";
